requirements.txt
flask==3.0.0
undetected-chromedriver==3.5.4
selenium==4.15.2
python-docx==1.1.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
beautifulsoup4==4.12.2

